fix: error when copying from recycle bin on SDK 31+ - fix NoSuchElementException, thrown when copying files from the recycle bin on Android 12+ devices. - only convert the paths to URIs needed for MediaStore.createWriteRequest  if the path is  not a recycle bin path - add extension function FileDirItem.isRecycleBinPath for checking if a FileDirItem's path starts with the recycleBinPath to prevent duplication

+++ commons/src/main/kotlin/com/simplemobiletools/commons/extensions/Activity.kt
@@ -724,16 +724,13 @@ fun BaseSimpleActivity.deleteFilesBg(files: List<FileDirItem>, allowDeleteFolder
                 return@checkManageMediaOrHandleSAFDialogSdk30
             }
 
-            if (canManageMedia()) {
+            val recycleBinPath = firstFile.isRecycleBinPath(this)
+            if (canManageMedia() && !recycleBinPath) {
                 val fileUris = getFileUrisFromFileDirItems(files).second
-                if (fileUris.size == files.size) {
-                    deleteSDK30Uris(fileUris) { success ->
-                        runOnUiThread {
-                            callback?.invoke(success)
-                        }
+                deleteSDK30Uris(fileUris) { success ->
+                    runOnUiThread {
+                        callback?.invoke(success)
                     }
-                } else {
-                    deleteFilesCasual(files, allowDeleteFolder, callback)
                 }
             } else {
                 deleteFilesCasual(files, allowDeleteFolder, callback)
